requirements.txt
# AI Security CLI (tool under evaluation)
ai-security-cli>=1.0.0b2

# Baseline tools
semgrep>=1.50.0
bandit>=1.7.0
detect-secrets>=1.4.0
pip-audit>=2.6.0

# Red team tools
garak>=0.9.0

# LLM providers for live testing
openai>=1.0.0
anthropic>=0.8.0

# Analysis and reporting
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Utilities
pyyaml>=6.0.0
rich>=13.0.0
click>=8.0.0
requests>=2.28.0
httpx>=0.24.0

# For scorecard (optional)
# scorecard-action (use via GitHub Actions)
